final updates
production of site has come to an end as per the client request further development was halted. the client at this time has no intention of using this site or setting up this service at the time but using the images and ideas that came from the site in a class project.